Split up servers component
